Javascript bindings for CEF 3 are almost ready, a bit of work is still required.
Fix to the compile.py script.

Updated the wxpython script, added js bindings.

Updated makefiles to display warnings and treat warnings
as errors.

+++ cefpython/cef3/linux/compile.py
@@ -76,11 +76,10 @@ if ret != 0:
     if what != "y":
         sys.exit(1)
 subprocess_exe = "./../linux/binaries_%s/subprocess" % (BITS)
-shutil.copyfile("./subprocess", subprocess_exe)
-st = os.stat(subprocess_exe)
-os.chmod(subprocess_exe, st.st_mode | stat.S_IEXEC)
-
-
+if os.path.exists("./subprocess"):
+    shutil.copyfile("./subprocess", subprocess_exe)
+    st = os.stat(subprocess_exe)
+    os.chmod(subprocess_exe, st.st_mode | stat.S_IEXEC)
 
 # os.chdir("./../v8function_handler/")
 # ret = subprocess.call("make -f Makefile", shell=True)
